refactor batch invariance folder
