+++ vllm/model_executor/determinism/batch_invariant_configs.py
