+++ vllm/model_executor/layers/layernorm.py
@@ -11,7 +11,7 @@ import vllm.kernels  # noqa: F401
 from vllm import envs, ir
 from vllm.logger import init_logger
 from vllm.model_executor.custom_op import CustomOp
-from vllm.model_executor.layers.batch_invariant import rms_norm_batch_invariant
+from vllm.model_executor.determinism.batch_invariant import rms_norm_batch_invariant
 
 logger = init_logger(__name__)
